Refactoring
Rename `internals` to `core`

Move `cmd` into domains

Move some packages from `core` to particular domains

Enable syslog only if `SYSLOG` env var is set
