+++ domains/api/state.go
@@ -11,13 +11,13 @@ import (
 
 	"github.com/gin-gonic/gin"
 	"github.com/outcatcher/anwil/domains/api/handlers"
-	"github.com/outcatcher/anwil/domains/internals/config"
-	configSchema "github.com/outcatcher/anwil/domains/internals/config/schema"
-	"github.com/outcatcher/anwil/domains/internals/logging"
-	"github.com/outcatcher/anwil/domains/internals/services"
-	svcSchema "github.com/outcatcher/anwil/domains/internals/services/schema"
-	"github.com/outcatcher/anwil/domains/internals/storage"
-	storageSchema "github.com/outcatcher/anwil/domains/internals/storage/schema"
+	"github.com/outcatcher/anwil/domains/core/config"
+	configSchema "github.com/outcatcher/anwil/domains/core/config/schema"
+	"github.com/outcatcher/anwil/domains/core/logging"
+	"github.com/outcatcher/anwil/domains/core/services"
+	svcSchema "github.com/outcatcher/anwil/domains/core/services/schema"
+	"github.com/outcatcher/anwil/domains/storage"
+	storageSchema "github.com/outcatcher/anwil/domains/storage/schema"
 	users "github.com/outcatcher/anwil/domains/users/service"
 	usersSchema "github.com/outcatcher/anwil/domains/users/service/schema"
 )
