+++ domains/core/doc.go
@@ -1,0 +1,4 @@
+/*
+Package core contains internal and core functionality domain
+*/
+package core
