+++ domains/core/logging/default_win.go
@@ -1,0 +1,10 @@
+//go:build windows
+
+package logging
+
+import "log"
+
+// GetDefaultLogger returns default stdout logger.
+func GetDefaultLogger() *log.Logger {
+	return log.Default()
+}
