+++ domains/core/logging/doc.go
