+++ domains/core/services/schema/errors.go
