--- domains/internals/doc.go
@@ -1,4 +1,0 @@
-/*
-Package internals contains internal and core functionality domain
-*/
-package internals
