+++ domains/storage/cmd/migrate/main.go
@@ -7,41 +7,38 @@ import (
 	"context"
 	"flag"
 	"log"
-	"log/syslog"
 	"path/filepath"
 
-	"github.com/outcatcher/anwil/domains/internals/config"
-	"github.com/outcatcher/anwil/domains/internals/storage"
+	"github.com/outcatcher/anwil/domains/core/config"
+	"github.com/outcatcher/anwil/domains/core/logging"
+	"github.com/outcatcher/anwil/domains/storage"
 )
 
 func main() {
-	sysLogger, err := syslog.New(syslog.LOG_INFO, "anwil-migrate")
-	if err != nil {
-		log.Fatalln(err)
-	}
+	logger := logging.GetDefaultLogger()
 
-	log.SetOutput(sysLogger)
+	log.SetOutput(logger.Writer()) // for subsequent goose calls
 
 	argConfigPath := flag.String("config", "", "Configuration path")
 	flag.Parse()
 
 	if *argConfigPath == "" {
-		log.Fatalf("please provide configuation path")
+		logger.Fatalf("please provide configuation path")
 	}
 
 	configPath, err := filepath.Abs(filepath.Clean(*argConfigPath))
 	if err != nil {
-		log.Fatal(err)
+		logger.Fatal(err)
 	}
 
 	ctx := context.Background()
 
 	cfg, err := config.LoadServerConfiguration(ctx, configPath)
 	if err != nil {
-		log.Fatal(err)
+		logger.Fatal(err)
 	}
 
 	if err := storage.ApplyMigrations(cfg.DB); err != nil {
-		log.Fatal(err)
+		logger.Fatal(err)
 	}
 }
