+++ domains/users/handlers/handlers.go
@@ -5,8 +5,8 @@ package handlers
 
 import (
 	"github.com/gin-gonic/gin"
-	services "github.com/outcatcher/anwil/domains/internals/services/schema"
-	"github.com/outcatcher/anwil/domains/internals/validation"
+	services "github.com/outcatcher/anwil/domains/core/services/schema"
+	"github.com/outcatcher/anwil/domains/core/validation"
 	"github.com/outcatcher/anwil/domains/users/service/schema"
 )
