+++ migrate.Dockerfile
@@ -4,14 +4,14 @@ WORKDIR /opt/build
 
 COPY . .
 
-RUN go build -trimpath -o ./migrate ./cmd/migrate/main.go
+RUN go build -trimpath -o ./migrate ./domains/storage/cmd/migrate/main.go
 
 FROM alpine:3.17.2
 
 WORKDIR /opt/migrate
 
 COPY --from=builder /opt/build/migrate ./migrate
-COPY ./migrations ./migrations
+COPY migrations ./migrations
 COPY ./anwil-config.yaml ./config.yaml
 
 CMD ["./migrate", "-config", "./config.yaml"]
