+++ testing/suite_test.go
@@ -24,10 +24,10 @@ import (
 	"github.com/docker/go-connections/nat"
 	"github.com/gin-gonic/gin"
 	"github.com/outcatcher/anwil/domains/api"
-	"github.com/outcatcher/anwil/domains/internals/config"
-	"github.com/outcatcher/anwil/domains/internals/config/schema"
-	"github.com/outcatcher/anwil/domains/internals/logging"
-	"github.com/outcatcher/anwil/domains/internals/storage"
+	"github.com/outcatcher/anwil/domains/core/config"
+	"github.com/outcatcher/anwil/domains/core/config/schema"
+	"github.com/outcatcher/anwil/domains/core/logging"
+	"github.com/outcatcher/anwil/domains/storage"
 	usersDTO "github.com/outcatcher/anwil/domains/users/dto"
 	usersSchema "github.com/outcatcher/anwil/domains/users/service/schema"
 	"github.com/stretchr/testify/require"
